// src/models/Lead.ts
export interface Lead {
  lead_id: number;
  name: string;
  surname: string;
  phone: string;
  mail: string;
  town: string;
  post_code: string;
  adress: string;
  client_type: string;
  age: string;
  type: string;
  campaign: string;
  product: string;
  campaign_image: string;
  price: number;
  region: string;
  area: string;
  owner: string;
  status: string
}